META-2144: Remove kv_api_sync.rs
